+++ client/package.json
@@ -18,6 +18,7 @@
     "react": "^19.2.5",
     "react-dom": "^19.2.5",
     "react-i18next": "^17.0.8",
+    "react-router-dom": "^7.16.0",
     "tailwindcss": "^4.2.4",
     "uuid": "^14.0.0"
   },
